stm32f0: adjust wwdg threshold signatures to support 12 bit resolution

+++ include/libopencm3/stm32/f0/adc.h
@@ -190,8 +190,8 @@ void adc_calibrate_wait_finish(uint32_t adc)
 void adc_enable_analog_watchdog_on_all_channels(uint32_t adc);
 void adc_enable_analog_watchdog_on_selected_channel(uint32_t adc, uint8_t chan);
 void adc_disable_analog_watchdog(uint32_t adc);
-void adc_set_watchdog_high_threshold(uint32_t adc, uint8_t threshold);
-void adc_set_watchdog_low_threshold(uint32_t adc, uint8_t threshold);
+void adc_set_watchdog_high_threshold(uint32_t adc, uint16_t threshold);
+void adc_set_watchdog_low_threshold(uint32_t adc, uint16_t threshold);
 
 END_DECLS
